Refactor OperationDescriptor and *Selector types 1/n: Add RequestDescriptor alias
Summary:
## Context

The goal of this stack is to refactor the  OperationDescriptor and `-Selector` types to make the structure more directly represent how the parts of these values are used, avoid cycles, and avoid destructuring/restructuring. The goal is to get the  point where Environment::lookup() accepts a single "OwnedSelector" argument - which can be singular, plural, or a new connection variant - and returns a corresponding (input-dependent) result type.

## This Diff

As a very first step, this creates a new `RequestDescriptor` type that has the same shape as `OperationDescriptor` but  where `fragment` and  `root` are mixed. This a) sets up the new type and b) confirms that the fragment/root properties aren't actually consumed by code that consumes fragment owners.

Reviewed By: kassens

Differential Revision: D16502120

fbshipit-source-id: 61ecde64a8809638a052e123cce29b3ea552ef2b