simple optimization for validateRelayServerOnlyDirectives
Summary:
We should see if we can move this validation earlier before we inline all fragments. This would make the validation rules a bit more complicated as it would have to traverse fragments, but a lot faster.

This is a simple optimization that cuts the time a bunch by no looking at the full arguments and other fields of linked and scalar fields (by far the most common AST nodes)

Reviewed By: tyao1

Differential Revision: D17046293

fbshipit-source-id: f0f475e1e55cd366dcd25f55b84b7270618136b5