DataChecker: ignore missing id on non-Node records
Summary:
Relay doesn't currently store information about which concrete types implement which interfaces. Instead, in the normalization AST selections on abstract types are flattened into their parents. One consequence of this is that if a field sometimes returns a Node but not always, the compiler will first generate  a `... on Node { id }` selection  and then later flatten it to just `id`.  When DataChecker runs, if the field value happens to be a type not implementing  Node then `id` will appear missing.

A general fix to this for arbitrary interfaces is more involved - runtime type introspection as is required to check  if a type implements an interface is not free and different programming languages make different tradeoffs here (for example this is opt-in in Rust). This diff fixes the special case of `...  on Node { id }` by not considering id to be missing on client records.

Reviewed By: kassens

Differential Revision: D17101169

fbshipit-source-id: 60883c07c643dce3b9f6091d2a8ca262453023a0