New Connections: more test cases, restore both snapshots and events during rebase/revert
Summary:
Adds more test cases for the new connection model, including multiple optimistic updates with revert as well as optimistic deletion of edges/nodes. This is combined with a more sophisticated reducer that prunes edges that are null or have a null node in order to ensure that reverting changes works as expected.

The approach still requires an O(n) snapshot of the connection state - now including both the subscription state and the events map - but my plan is to optimize this in a follow-up once all use-cases are covered and tested.

Reviewed By: kassens

Differential Revision: D17197732

fbshipit-source-id: 36cf3ecc626ae2f9e2bc972daac1d4af874aa2ba