New connections: revert subscriptions established while optimistic updates were applied
Summary: The previous mechanism for restoring connection subscription states only worked for subscriptions established prior to the creation of the snapshot: any subscriptions established afterwards (ie while there were optimistic updates applied) wouldn't revert correctly since there was nothing to revert to. The fix is to detect this case, and recompute such subscriptions from scratch against the restored base state of the store (note that this is sensitive to the backup RecordSource having been published first).

Reviewed By: jstejada

Differential Revision: D17216207

fbshipit-source-id: a34aa62af04f82f9d4926ef133e7be07ebfb4904