Make PublishQueue-test more agnostic of implementation details
Summary: Changes PublishQueue-test to be more agnostic of the implementation details. In particular, the intent of the tests was to assert the state of the store as viewed from the outside, but this was implemented in the test by checking for modifications to the base RecordSource. This is now changed to use getSource(), allowing the implementation of reverting/rebasing to be achieved via different approaches than the current backup mechanism.

Reviewed By: kassens

Differential Revision: D17267961

fbshipit-source-id: e3d57ff2b5f91f5ed5f6b35717069152dd3c0f08