Use connection snapshot/restore mechanism for regular subscriptions
Summary:
Changes PublishQueue/Store to use a new approach for optimistic updates. Previously, PublishQueue built a backup RecordSource that contained the old version of any data that was optimistically modified. When optimistic updates needed to be reverted or rebased, this backup would be published (just like any other change). This worked well prior to adding support for connection subscriptions, which build up derived state from the store. Reverting derived state requires restoring to a "final" (non-optimistic) snapshot of the derived state - which is mostly easy. However, publishing a backup source meant that we would have to unnecessarily process "update" events on connection subscriptions to account for rollbacks to edge data changes, only to then write over those updated states with the backup derived data.

The solution implemented here is to apply the same snapshot/restore mechanism used for connections to regular subscriptions too. Instead of creating a backup RecordSource, PublishQueue now tells the store to snapshot itself (prior to applying the first optimistic update) and later restore its state from that snapshot (when it needs to revert/rebase optimistic updates). The Store now keeps an underlying RecordSource that represents the final state as well as a layered OptimisticRecordSource that stacks optimistic changes on top of the final state. It also keeps both a "latest" snapshot as well as a backup snapshot for each subscription. Restoring final state only requires discarding the layered optimistic store and resetting any subscriptions whose latest state diverged from the backup state.

Reviewed By: kassens

Differential Revision: D17267960

fbshipit-source-id: b7413f4e4d39adc8dd484ad1fb0c8d3fd9c587df