Use global state for persisted queries to fix --watch (#2629)
Summary:
This is my attempt to fix the issue raised here https://github.com/facebook/relay/issues/2625

I'm using a map keyed by filename and ensuring all queries are written even if unchanged while watching. The map is initialised in scope when writing, so old files are omitted if removed and the state is therefore hopefully correct for each write. My assumption is there is a 1 to 1 relationship of filenames to query id.

The main logic is making sure persistQuery is called even if the query is unchanged (ie. the hashes match).
Pull Request resolved: https://github.com/facebook/relay/pull/2629

Reviewed By: jstejada, kassens

Differential Revision: D13946383

Pulled By: josephsavona

fbshipit-source-id: c66d052bf559111011eeddb6ee2ee98cb190cd6d