add event based log system to environment
Summary:
This is a start for a more unified instrumentation framework for Relay. The idea is that the RelayEnvironment has a log function that receives various named events. These events can be tied together with a `transactionID`, but it's up to the logging consumer to put these time spans together. This is different from some of the previous logging.

This adds a `environment.log(event)` API where the event is a union type of different events.
Associated events have a common `transactionID`. It's up to the logging consumer to re-construct the timeline or just ignore parts. This different from our current network logger which first collects all information of a request and then logs them in aggregate. The event based system makes it easier to add events from more parts of the stack and consume them as a stream while they're happening, such as displaying all subscriptions that are in flight.

To support making the logging ReactContext aware, we think we can allow setting an additional logging context that updates the Environment for a subtree — copying most properties, but adding an additional logger. This isn't implemented yet as we need to ensure mappings from environment to values such as the query resource don't relay on environment object identity, maybe instances would have a conceptional "identity key".

The one case that's not covered is the synchronous logging. I don't think this is active at the moment or if it is, we're not looking at it. Issuing an extra log event for every Store.lookup seems wasteful as this is an extremely common operation. We can keep the synchronous version of RelayProfiler logging around for now, but might want to just delete it and only instrument methods for the time we actively want to profile them from production. For everyday use, the revived mbench micro-benchmark could be useful.

Reviewed By: josephsavona

Differential Revision: D17454433

fbshipit-source-id: 1f46f12e35a4de5b44c6653c710409761801dd64