Allow pagination even if has_next is false
Summary:
This will give product developers more flexibility over when to send a pagination request, and there are legitimate cases where we'd still want to try to send a pagination request even if has_next_page was false. For example, new items might have been added to the connection after we received has_next as false.

The product code can manually check hasNext to decide to bail out of pagination

Reviewed By: kassens

Differential Revision: D17555629

fbshipit-source-id: 89f333a38b3c317f72e52b138e96e29a0e8ea4bc