Don't suspend indefinitely even if network response still contains missing data
Summary:
Currently, when a query suspends on a network request and  we receive a payload from the server, we only resolve that query if there is no missing data (for the query's fragment) on the returned payload, under the assumption that the payload would fetch everything that the query needs.

However, this assumption is incorrect, because there are cases in which fetching a query won't return data for all of the selections in the query; for example, when making selections on abstract types, if the actual concrete type doesn't implement the abstract type, we'll still have a selection that will always be missing, since we can't check schema information at runtime (T53445119).

Given that this can happen, if we don't resolve the query after the request completes, it will stay suspended indefinitely. Instead, we can always resolve the query; in fact this check is completely unecessary because when reading the fragment for the query (or descendant fragments), if the fragment has missing data and the request is still in flight, the fragment will suspend.

Reviewed By: josephsavona

Differential Revision: D17573316

fbshipit-source-id: e2e6ae57c33512654c1b150689851747026d776a