Correctly dispose of requests that are still in flight when unmounting
Summary:
Currently, we relied solely on a React ref to cancel network requests in flight when the query changed or a query component unmounts.

However, with Suspense, when a component commits after it was suspended (e.g. waiting on a query), the component will be reset, meaning the ref will also be reset, causing us to lose a reference to the network subscription; by the time the component unmounts, we don't have a reference to the network subscription that we can use to cancel a request.

This usually wasn't a problem because when the component unsuspends and commits it usually means that the network request has completed, but there is the the possibility to have long lived requests that will cause the query to mount on the initial payload, and which we want to cancel if the query unmounts.

The fix here is to keep a reference to the network subscription /outside/ the component, in `QueryResource`, like we do with for retaining the query, which knows how to deal with the nature of Suspense and the fact render can be retried multiple times and that only on commit we can successfully hold on to resources.

Reviewed By: kassens

Differential Revision: D17487063

fbshipit-source-id: 2e0f4d2a007c9a527694397fad1db8ae99ab8fca