Ensure that canceling a request resolves/completes promises or observables observing if request is in flight
Summary:
Currently, fetchQueryInternal will dedupe identical requests in flight and use a ReplaySubject to replay all events that have occurred to new subscribers that attempt to start a network request that is already in flight.

Additionally, we currently expose some helpers to observe whether a request is in flight: `getPromiseForRequestInFlight` and `getObservableForRequestInFlight`. However, if the underlying network request is canceled, the promise/observable won't resolve and stay pending indefinitely (see also summary in D17512468).

This commit fixes this by making sure that the unsubscribe event is propagated to the promise/observables waiting on the pending request, and resolving them when that occurs.

Reviewed By: josephsavona

Differential Revision: D17528534

fbshipit-source-id: 319356ad863b0a9bafa31221a5551ff5a5375ba7