Allow selecting __id anywhere
Summary:
Allows selecting `__id` anywhere that `__typename` can be selected: on objects, interfaces, *or* unions. The current motivating case is new connections, where it's convenient to be able to access the cache id of both edges and edges.node records (our ConnectionHandler approach relies on this too, but the implementation there is an Updater function and can just call getDataID() on the edges/node proxy objects - new connections use raw GraphQL to access the data they need, hence we need a way to translate that getDataID() call into a valid query).

There's another case that we've also considered, which is making it easier to determine the Relay cache id of an object in order to execute mutations against it. For example, it's currently difficult to delete a client record, since `RecordSourceProxy::delete()` takes an id, and there' no way to select one for a client record via a fragment. Today that use-case has to figure out the id via the nearest parent record with an id, but if `__id` can be selected it's trivial (`store.delete(record.__id)` always works for any record).

Note that this includes a fix to our `isServerDefinedField` logic, which happened to be incorrect for fields on unions (found by testing that `__id` on a union type is stripped in the print query.)

Reviewed By: kassens

Differential Revision: D17607210

fbshipit-source-id: c3b0ab4303cb17725088b45b0bede96076847422