New connections: support connections where node type is a union
Summary: Adds support for connections where `edges.node` is a union type: primarily this means a) allowing `__id` selections on union types (done in the previous diff in the stack) and b) adding an or case to allow the edge type to be a union. The rest is just tests.

Reviewed By: alunyov

Differential Revision: D17608770

fbshipit-source-id: c29727bbd1f889b4b7b7726c009fa8d76b22b7ac