Fix @refetchable connection metadata for custom handlers
Summary: Fixes an issue where `refetchable` doesn't generate connection metadata correctly when `connection` or `stream_connection` are used with custom handlers. The metadata extraction logic looked for fields with a handler with name="connection" or for the connection/stream_connection directives, but those directives are stripped in ConnectionTransform. The fix here is to not strip the directives immediately in ConnectionTransform so that they're still visible to RefetchableTransform - and then rely on the print context stripping client directive extensions anyway.

Reviewed By: jstejada

Differential Revision: D17612983

fbshipit-source-id: 7de2023affb25a937ae35130039d47a7a43e1e0f