Relay Schema Interface
Summary:
### Context

This diff introduces the new GraphQL schema interface for Relay Compiler.

The main motivation for this change - the work on the performance improvements in the compiler.
In the relay compiler, a big amount of memory is reserved for the `graphql-js` representation of the schema.

To reduce the memory usage in the compiler we need to find a light-weight representation of the schema.

But, since the core compiler primitives (RelayParser, Transforms, Code Generation) are coupled with
the `graphql-js` implementation of the schema, there is no easy way to optimize the performance there.

So the plan is to refactor all the places where we're using `graphql-js` classes and functions to use
Relay Compiler Schema Interfaces instead.

This diff has an initial implementation of such interface.
The next diff in the stack - is the actual refactoring of the compiler.

### A high-level API overview:

The core elements of the Schema API.

To create an Instance of the `Schema`, the module exposes a function `create` which receives a
`Source` of the schema SDL file.

The main method to get a type for the schema is:

  schema.getTypeFromString(typeName: string): ?TypeID

where, TypeID - is an opaque unique identifier of the GraphQL type.

This method will return typeID if GraphQL type is available in the server (or client) version of the schema.

There also `expectSomething(...)` methods avialable

  schema.expectTypeFromString(typeName, message?, localtions?, nodes?): TypeID

This method will throw an exception, if type is not foind.

To get a field information:

  schema.getFieldByName(typeID: TypeID, fieldName: string): ?FieldID

There also strict method:
  schema.expectField(typeID, fieldName): FieldID

With `fieldID` we can request field configuration parameters:
  schema.getFieldConfig(fieldID: FieldID): FieldConfiguration // type, arguments

there also a set of different `isSomething(typeID)` and `assertSomthing(typeID)` methods - that should perform different checks: if a type is null, list, a specific type, an interface, etc...

Reviewed By: josephsavona

Differential Revision: D17419703

fbshipit-source-id: 87627c4cf40b519d0f9f2fe98c63ad8c0eb1294d